requirements.txt
google-genai==1.16.0
groq==0.13.1
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
faster-whisper==1.1.1
yt-dlp==2025.1.26
edge-tts==7.0.2
opencv-python-headless==4.10.0.84
PyYAML==6.0.2
pytz==2024.2
requests==2.32.3
pydantic==2.9.2
numpy==1.26.4
